Fix regression since JW's changes
Ignore-this: e370e2e691773bb88f355f6d8ebb103

This was introduced when merging the current head with changes Jason sent via
email. I am unsure as to which implementation is "correct", but this one gives the
same result as when Jason last inspected the code.

This change means that the total mass of the peptide is no longer part of the
ion series mass ladders...

darcs-hash:20090708052917-9241b-4321b7492eaec79acf0e9e5109a86310852acda5.gz