Speed up spectrum code, fix region normalising bug
Ignore-this: e4bbdcf7049992e45f2a433b1d56d97e

The spectrum code is now only as large as necessary, specifically the input
range plus 75 bins on each end to account for the FFT "wings".

When normalising the values in the 10 equally sized windows, values that fell
beyond the range of the last window were used to find the maximum value of that
window, and only set to zero in the second normalisation pass. This has been
corrected.

darcs-hash:20090708233422-9241b-7e3a539775471e9dcce795e60fe013e06a553435.gz